Added new endpoints to download and save image

+++ webpeditor_app/jobs/daily/check_session_token.py
@@ -7,10 +7,9 @@ from django.utils import timezone
 from django_extensions.management.jobs import DailyJob
 
 from webpeditor import settings
-from webpeditor_app.models.database.models import OriginalImage
 from webpeditor_app.services.image_services.folder_service import delete_folder_by_expiry
 from webpeditor_app.services.image_services.image_service import delete_old_image_in_db_and_local, \
-    get_serialized_data_original_image, get_serialized_data_edited_image
+    get_serialized_data_original_image, get_serialized_data_edited_image, get_all_original_images, get_all_edited_images
 
 logging.basicConfig(level=logging.INFO)
 
@@ -22,15 +21,17 @@ class Job(DailyJob):
         deserialized_data_original_image = get_serialized_data_original_image()
         deserialized_data_edited_image = get_serialized_data_edited_image()
 
-        try:
-            original_images = OriginalImage.objects.all()
-        except OriginalImage.DoesNotExist as error:
-            raise error
+        original_images = get_all_original_images()
+        edited_images = get_all_edited_images()
 
         media_root: Path = settings.MEDIA_ROOT
 
-        logging.info(f"\n--- Original Image object(s) in db ---\n{json.dumps(deserialized_data_original_image, indent=4)}")
-        logging.info(f"\n--- Edited Image object(s) in db ---\n{json.dumps(deserialized_data_edited_image, indent=4)}")
+        logging.info(
+            f"\n--- Original Image object(s) in db ---\n{json.dumps(deserialized_data_original_image, indent=4)}"
+        )
+        logging.info(
+            f"\n--- Edited Image object(s) in db ---\n{json.dumps(deserialized_data_edited_image, indent=4)}"
+        )
 
         counter = 0
         if len(original_images) > 0:
@@ -38,17 +39,26 @@ class Job(DailyJob):
                 user_id = image.user_id
                 session_id_expiration_date = image.session_id_expiration_date
 
-                if user_id and (timezone.now() > session_id_expiration_date):
+                if timezone.now() > session_id_expiration_date:
                     delete_old_image_in_db_and_local(user_id)
                     counter += 1
 
-                if not user_id:
+                if user_id and not (user_id in os.listdir(media_root)):
                     delete_folder_by_expiry(media_root)
 
-                if user_id and not (user_id in os.listdir(media_root)):
-                    delete_old_image_in_db_and_local(user_id)
+                if not user_id:
+                    logging.error("There is no original images. Deleting users folders...")
+                    delete_folder_by_expiry(media_root)
 
             logging.info(f"Deleted collections in db: {counter}")
 
-        elif len(deserialized_data_original_image) == 0 or len(deserialized_data_edited_image) == 0:
+        if len(edited_images) > 0:
+            for image in edited_images:
+                user_id = image.user_id
+                if not user_id:
+                    delete_folder_by_expiry(media_root)
+
+                delete_old_image_in_db_and_local(user_id)
+
+        if len(deserialized_data_original_image) == 0 or len(deserialized_data_edited_image) == 0:
             delete_folder_by_expiry(media_root)
